sale_order_import: re-organize test files
