+++ sale_order_import/tests/__init__.py
@@ -2,3 +2,4 @@
 
 from . import test_sale_order
 from . import test_order_import
+from . import test_parse_validate
